# internal/openapi/openapi.yaml
...
openapi: 3.1.0
info:
  title: GSE Terminal API
  version: "1.0.0"
  summary: Programmatic access to the Ghana Stock Exchange wire.
  description: |
    Every figure on the GSE Terminal — each tick, close, sentiment score
    and sector aggregate — is available over HTTPS as JSON under the
    `/v1` path.

    ## Authentication

    Most endpoints require a bearer token in the `Authorization` header:

    ```
    Authorization: Bearer ges_live_...
    ```

    Keys are provisioned from **Terminal → Account → API keys** and are
    shown only at creation time. A session cookie works equivalently when
    the call originates from a browser signed into the Terminal.

    ## Rate budget

    - 60 requests per minute, per IP, across the whole surface.
    - LLM endpoints (`/v1/ai-insight`, `/v1/query`) are additionally
      capped at 10 requests per minute, per authenticated user.
    - Exceeding either budget returns `429` with an exponential-backoff
      hint in the response body.

    ## Errors

    Every error response is a JSON body with a single `error` field.
    The HTTP status is authoritative.
  contact:
    name: GSE Terminal
    url: https://your-host/developers
  license:
    name: Proprietary
    url: https://your-host/terms
servers:
  - url: https://{host}/v1
    description: Production
    variables:
      host:
        default: your-host
        description: Your deployment host.
  - url: http://localhost:8080/v1
    description: Local development
...
security:
  - bearerApiKey: []
  - sessionCookie: []

# ──────────────────────────────────────────────────────────────────────────
# Paths
# ──────────────────────────────────────────────────────────────────────────
paths:
...
  /watchlist:
    get:
      tags: [User]
      operationId: getWatchlist
      summary: Caller's watchlist
      description: Starred symbols with the latest market snapshot joined on.
      responses:
        '200':
          description: Watchlist envelope.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WatchlistResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [User]
      operationId: toggleWatchlist
      summary: Add or remove a symbol from the watchlist
      description: Idempotent toggle — posting the same symbol twice returns it to its previous state.
      parameters:
        - name: symbol
          in: query
          required: true
          description: Ticker to toggle.
          schema: { type: string, example: MTNGH }
      responses:
        '200':
          description: New membership state.
          content:
            application/json:
              schema:
                type: object
                required: [isWatchlisted]
                properties:
                  isWatchlisted: { type: boolean }
                example: { isWatchlisted: true }
        '401': { $ref: '#/components/responses/Unauthorized' }
...
  /me/api-keys:
    get:
      tags: [Keys]
      operationId: listAPIKeys
      summary: List your API keys
      description: Only the prefix is stored server-side. The raw key is returned only at creation time.
      responses:
        '200':
          description: Array of API keys (no raw secret).
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/APIKey' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    post:
      tags: [Keys]
      operationId: createAPIKey
      summary: Provision a new API key
      description: The raw `key` field is returned **once**. Store it at the callsite — it cannot be recovered later.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Display label, max 200 characters.
                  example: screener-bot
      responses:
        '200':
          description: The new key — including the raw secret.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/APIKeyWithSecret' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /me/api-keys/{id}:
    delete:
      tags: [Keys]
      operationId: revokeAPIKey
      summary: Revoke an API key
      description: Subsequent requests presenting this key return `401`.
      parameters:
        - name: id
          in: path
          required: true
          description: Key ID.
          schema: { type: integer, format: int64, example: 3 }
      responses:
        '200':
          description: Revocation confirmation.
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, example: revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }

# ──────────────────────────────────────────────────────────────────────────
# Components
# ──────────────────────────────────────────────────────────────────────────
components:
  securitySchemes:
    bearerApiKey:
      type: http
      scheme: bearer
      bearerFormat: ges_live_*
      description: |
        Provision a key from **Terminal → Account → API keys**. The raw
        secret is shown only at creation time.
    sessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: JWT-backed session cookie issued by `/login`. Equivalent to a bearer key for browser-originated calls.
  responses:
...
    Unauthorized:
      description: Missing or invalid bearer token / session cookie.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "unauthorized" }
...
    NotFound:
      description: No such resource, or not owned by the caller.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { error: "not found" }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string, description: Human-readable error message. }
...
    MarketSnapshot:
      type: object
      required: [symbol, lastPrice]
      properties:
        symbol:        { type: string, example: MTNGH }
        openPrice:     { type: number, example: 1.50 }
        lastPrice:     { type: number, example: 1.54 }
        priceChange:   { type: number, example: 0.04 }
        percentChange: { type: number, example: 2.67 }
        volume:        { type: integer, format: int64, example: 211800 }
...
    WatchlistResponse:
      type: object
      required: [symbols]
      properties:
        symbols:
          type: array
          items: { type: string }
          example: [MTNGH, GCB, FML]
        details:
          type: array
          items: { $ref: '#/components/schemas/MarketSnapshot' }
...
    APIKey:
      type: object
      required: [id, prefix]
      properties:
        id:         { type: integer, format: int64, example: 3 }
        prefix:     { type: string, description: First few characters of the key, shown for recognition., example: 'ges_live_aZ8k' }
        name:       { type: string, example: my-laptop }
        createdAt:  { type: string, format: date-time, example: '2026-03-22T10:12:00Z' }
        lastUsedAt: { type: string, format: date-time, nullable: true, example: '2026-04-14T09:40:00Z' }
    APIKeyWithSecret:
      allOf:
        - $ref: '#/components/schemas/APIKey'
        - type: object
          required: [key]
          properties:
            key:
              type: string
              description: The raw secret. Returned **only** from this endpoint, **only** at creation time.
              example: ges_live_qR4v9n8xK2pLm1dC5wY3zH7vE6tB0sJ4uA8iO